Week2/Day5/health_tracker/requirements.txt
Flask
psycopg2-binary
Faker
requests
matplotlib
plotly
python-dotenv
